+++ src/main/java/com/example/take_project/rests/ClientPackageREST.java
@@ -5,6 +5,7 @@ import com.example.take_project.services.ClientPackageServiceInterface;
 import jakarta.ejb.EJB;
 import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
 
 import java.util.List;
 
@@ -23,26 +24,33 @@ public class ClientPackageREST {
 
     @GET
     @Path("/id/{id}")
-    public ClientPackage getById(@PathParam("id") Long id){
-        return clientPackageService.getById(id);
+    public Response getById(@PathParam("id") Long id){
+        ClientPackage cp = clientPackageService.getById(id);
+        if (cp == null) return Response.status(Response.Status.NOT_FOUND).build();
+        return Response.ok(cp).build();
     }
 
     @POST
     @Path("/addnew")
-    public void addNew(ClientPackage clientPackage){
+    public Response addNew(ClientPackage clientPackage){
         clientPackage.setId(null);
         clientPackageService.addNew(clientPackage);
+        return Response.ok().build();
     }
 
     @DELETE
     @Path("/delete/{id}")
-    public void delete(@PathParam("id") Long id){
+    public Response delete(@PathParam("id") Long id){
+        if (!clientPackageService.checkIfEntityWithIdExists(id)) return Response.status(Response.Status.NOT_FOUND).build();
         clientPackageService.delete(id);
+        return Response.ok().build();
     }
 
     @PUT
     @Path("/update")
-    public void update(ClientPackage cp){
+    public Response update(ClientPackage cp){
+        if (!clientPackageService.checkIfEntityWithIdExists(cp.getId())) return Response.status(Response.Status.NOT_FOUND).build();
         clientPackageService.update(cp);
+        return Response.ok().build();
     }
 }
